Some clean-up of unused variables.
Now determines default present list before replacing <name> with
<scribe>.

Now always guesses the same scribe if there are two or more
candidates, thanks to sorting the names. (Not important in itself, but
makes regression testing possible.)

Fix error in previous check-in: error in "Meeting:" regexp.

Removed unusued function ProcessRalphURL.

Now actually does what the previous check-in comment said:
s/@@/http://example/foo/ treats the last "/" as a delimiter.

Fixes ambiguity when a "]" follows a URL. And also removes white space
after a URL if it is probably only there to delimit the URL.

Recognize https URLs.

Use lc instead of tr/A-Z/a-z/ and also use it in more places so that
more names that differ only in case are reocgnized.